Remove failing unit test.
This removes a unit test for WriteArchiveAsync. The test expected an existing archive file to exist to source item data from. The test was not providing this data. When writing the archive, NefsLib would try to read past the end of the source archive to copy this non-existent data into the new archive.